vnet_tracker/requirements.txt
azure-functions
azure-identity
azure-mgmt-network
azure-cosmos